refactor(server-rs): split into the `sipi` library + a `cli-rs` binary
Realise decision #9's library boundary physically (plan 02 §7.5 M0): the
monolithic server-rs crate becomes the `sipi` library (//src/server-rs:lib,
crate `sipi`) exposing `run(config, overrides, drain_timeout)` + a Rust-native
`ServerOverrides`; the default binary moves to a new `cli-rs` crate
(//src/cli-rs:sipi) that owns `main`, the argv→verb dispatch, the `server`
verb's clap, and its own `extern "C" sipi_cli_main` binding.

The two FFI surfaces stay disjoint: server-rs keeps the server seam
(//src/ffi:sipi_ffi → //src:engine, which is why it drops the //src/cli:cli_app
dep); cli-rs binds only sipi_cli_main and adds //src/cli:cli_app for the offline
verbs. The engine links once into the final binary.

cli-rs's third-party deps resolve from a new @crates_cli crate_universe repo (a
separate single-manifest from_cargo, mirroring the existing crates/crates_prod
split) — a multi-manifest crates_prod broke all_crate_deps' in-repo package-path
resolution. The cli-rs → server-rs:lib edge is a Bazel deps edge only, no Cargo
path dep.

Retargets SIPI_BIN → //src/cli-rs:sipi, bazel-build-server + run-otel to the new
binary, and updates the stale e2e docstrings + CLAUDE.md.

Structural, no behaviour change. Verified: build green; all_e2e holds the 13/23
baseline (no regressions); unit + cli/cli_json/health/smoke/server-verb paths
pass through the new binary.
