+++ test/e2e/sipi_e2e_test.bzl
@@ -9,7 +9,7 @@ others aren't.
 What the macro injects:
 
   * `data`:
-      - `//src/cli:sipi`             sipi binary under test, via `SIPI_BIN`.
+      - `//src/cli-rs:sipi`      Rust shell binary under test, via `SIPI_BIN`.
       - `//:test_fixtures`       `test/_test_data/`, `config/`,
                                  `scripts/`, `server/` materialised so
                                  `sipi_e2e::repo_root()` resolves under
@@ -20,7 +20,7 @@ What the macro injects:
                                  `tests/snapshots/`.
 
   * `env`:
-      - `SIPI_BIN`               `$(rootpath //src/cli:sipi)` — runfiles-
+      - `SIPI_BIN`               `$(rootpath //src/cli-rs:sipi)` — runfiles-
                                  relative; `SipiServer` canonicalises
                                  to absolute before spawning.
       - `SIPI_REPO_ROOT`         `$(rootpath //:test_fixtures)` — the
@@ -109,7 +109,7 @@ def sipi_e2e_test(
         # only a couple of tests consume it but the cost is trivial and
         # makes adding new `assert_*_snapshot!()` calls a no-op wiring-wise.
         data = [
-            "//src/server-rs:sipi_server",
+            "//src/cli-rs:sipi",
             "//:test_fixtures",
             "//:lsan_suppressions",
             ":snapshots",
@@ -119,7 +119,7 @@ def sipi_e2e_test(
             # serves `server` natively and forwards every offline subcommand to
             # the C++ CLI via sipi_cli_main, so it covers both the server and CLI
             # e2e suites. The C++ `//src/cli:sipi` server is retired at the delete.
-            "SIPI_BIN": "$(rootpath //src/server-rs:sipi_server)",
+            "SIPI_BIN": "$(rootpath //src/cli-rs:sipi)",
             # Points at the `copy_to_directory` output that materialises
             # `version.txt`, `test/_test_data/`, `config/`, `scripts/`,
             # `server/` as real files (no symlinks). Required so sipi's
